Added hardcoded bootstrap version - RFC
I have added the version number that a created bootstrap file will get to this script.
This means that we will have to alter this number for each new bootstrapper we want created.
I think this is just fine because we must of course change bootstrap.tar.gz each time
to contain the latest faxien code we want present in a bootstrapper. We will now have
to alter the version info along with that each time.  Seems like a good idea to me. Dave
Peticolas first submitted a patch for something like this a month or so ago.  I am just
coming around ;-)

Secondly - and not really looking for comment here although it is always welcome -
I have added new text prompting people to enter adhoc arch specific info to a
bootstrapper.